Feature/encrypt plugin build

+++ sdk/utils/fs/rm-empty.go
@@ -12,10 +12,12 @@ import (
 	"path/filepath"
 )
 
-// Remove all contents of a directory without removing the directory itself. If the directory does not exist, create it.
+// Makes sure the directory exists and is empty. Removes contents of the directory if it already exists.
 func EmptyDir(dirPath string) error {
-	if err := os.RemoveAll(dirPath); err != nil {
-		return err
+	if Exists(dirPath) {
+		if err := os.RemoveAll(dirPath); err != nil {
+			return err
+		}
 	}
 	return os.MkdirAll(dirPath, PermDir)
 }
